+++ packages/excalidraw/index.tsx
@@ -10,7 +10,7 @@ import {
   applyDarkModeFilter,
   DEFAULT_IMAGE_OPTIONS,
   DEFAULT_UI_OPTIONS,
-  getStrokeWidthByKey,
+  getEffectiveStrokeWidth,
   isShallowEqual,
 } from "@excalidraw/common";
 
@@ -454,4 +454,4 @@ export function useExcalidrawStateValue(
 
 export { _useOnAppStateChange as useOnExcalidrawStateChange };
 
-export { applyDarkModeFilter, getStrokeWidthByKey };
+export { applyDarkModeFilter, getEffectiveStrokeWidth };